+++ examples/graphql/reinstall.sh
@@ -5,3 +5,4 @@ rm -rf node_modules
 rm -f rxdb-local.tgz
 npm run preinstall
 npm i
+npm run build